Minor cleanup
